requirements.txt
requests
paramiko>=3.1
qrcode[pil]
PySimpleGUI
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pyyaml>=6.0  # 可选，用于 ChatGPT 优化器生成 YAML 配置
